django/requirements.txt
Django==4.2
django-prometheus==2.3.1
python-logstash==0.4.8
python-json-logger==2.0.7
psycopg2-binary==2.9.5
requests==2.28.2
requests-cache==1.0.1
gunicorn
PyJWT>=2.8.0
pyotp>=2.9.0